save  work

+++ app/Repositories/Consultation/ConsultationDentisteRepository.php
@@ -127,6 +127,10 @@ class ConsultationDentisteRepository extends BaseRepository
     public function seanceUpdate($etat,$id){
         $seance = Seance::find($id);
         return $seance->update([
+            "date_consultation" => null,
+            "observation" => null,
+            "diagnostic" => null,
+            "bilan" => null,
             'etat' => $etat
         ]);
     }
@@ -141,21 +145,24 @@ class ConsultationDentisteRepository extends BaseRepository
         return $rendezVous;
     }
 
-    public function ConsultationModifier($input)
+    public function ConsultationModifier($input,$id)
     {
-        $consultation = Consultation::where('type',$input)->first();
-    
+        $consultation = Consultation::where('type', $input)->where('id', $id)->first();
+
         if ($consultation) {
-            Seance::where('consultation_id', $consultation->id)->delete();
-    
-            $rendezVousIds = Seance::where('consultation_id', $consultation->id)->pluck('rendezVous_id')->toArray();
-            RendezVous::whereIn('id', $rendezVousIds)->delete();
-    
+            $deletedSeances = Seance::where('consultation_id', $consultation->id)->delete();
+            
+            if ($deletedSeances === 0) {
+                $rendezVousIds = Seance::where('consultation_id', $consultation->id)->pluck('rendezVous_id')->toArray();
+                RendezVous::whereIn('id', $rendezVousIds)->delete();
+            }
+
             $consultation->update([
                 'etat' => Consultation::ETAT_EN_RENDEZVOUS
             ]);
         }
     }
+
     
 
     public function model(): string
